Use `requests` instead of `urllib.request`
For an undetermined reason, extracting the tar file of KooZic fails with
the following error: `EOFError: Compressed file ended before the end-of-
stream marker was reached`.

The root cause is not clear, since it is still working up to
`v1.1.0-beta.1`. However, `requests` seems to avoid this error.
Moreover, it has an easier API.