Minimum soft memory limit
A minimum soft memory limit of 256 Mo is too optimistic for real use
cases. Existing installations [1] with 1 Go of RAM and 1 CPU already
leads to a soft memory limit of 392 Mo. This is easily reached, leading
to systematic restart of the workers.

We pump the minimum value to 896 Mo. In practice, it is unlikely that
several workers will use this amount of RAM at the same time.

[1] https://github.com/DocMarty84/koozic/issues/36